create_function_working

+++ lambda/layers/python/shared/validators/amount_validator.py
@@ -1,7 +1,9 @@
+from decimal import Decimal
+
 def validate_amount(amount):
 
     try:
-        return float(amount)
+        return Decimal(amount)
 
     except (ValueError, TypeError):
         raise ValueError(
